+++ specs/lib/loadMtlSpec.js
@@ -343,6 +343,20 @@ describe('loadMtl', () => {
             expect(material.alphaMode).toBe('BLEND');
             expect(material.doubleSided).toBe(true);
         });
+
+        it('uses diffuse texture if diffuse and alpha are the same', () => {
+            options.metallicRoughness = true;
+
+            const material = loadMtl._createMaterial({
+                diffuseTexture : diffuseTexture,
+                alphaTexture : diffuseTexture
+            }, options);
+
+            const pbr = material.pbrMetallicRoughness;
+            expect(pbr.baseColorTexture).toBe(diffuseTexture);
+            expect(material.alphaMode).toBe('BLEND');
+            expect(material.doubleSided).toBe(true);
+        });
     });
 
     describe('specularGlossiness', () => {
@@ -435,6 +449,20 @@ describe('loadMtl', () => {
             expect(hasBlack).toBe(true);
             expect(hasWhite).toBe(true);
             expect(pbr.diffuseFactor[3]).toEqual(1);
+            expect(material.alphaMode).toBe('BLEND');
+            expect(material.doubleSided).toBe(true);
+        });
+
+        it('uses diffuse texture if diffuse and alpha are the same', () => {
+            options.specularGlossiness = true;
+
+            const material = loadMtl._createMaterial({
+                diffuseTexture : diffuseTexture,
+                alphaTexture : diffuseTexture
+            }, options);
+
+            const pbr = material.extensions.KHR_materials_pbrSpecularGlossiness;
+            expect(pbr.diffuseTexture).toEqual(diffuseTexture);
             expect(material.alphaMode).toBe('BLEND');
             expect(material.doubleSided).toBe(true);
         });